VI "Host - Get Local and ECAT Masters.vi" modified to solve bug #33 and implement a workaround for bug #15.
For "slow" modules, taking more time and throwing 65700 error, we exit right away and retry later with a maximum number of retries.
For error "65704", as CAR #729217 is long-time deferred, I propose following: detect 65704 error and if it happens. open a dialogue window to user, asking to do a deployment before redo an Auto-Detect or add manually missing cRIO modules in local chassis.