Fix EtherCAT FPGA deployment and add related tests (#93)
#91 broke without the development team knowing due to lack of testing. This fixes the broken feature and adds two different test strategies for ensuring it does not break again.

Testing the 9145 in FPGA mode required an additional change to the Scan Mode test: the target must be explicitly reverted to Scan Mode to clear the FPGA bitfile downloaded in the FPGA test. The FPGA download and revert code is borrowed from the System Explorer code, both of which have error handlers that may pop up if errors occur.

Additionally, the IP address of the target must be set in the System Definition before VS deploys. The IP address must be used instead of the hostname as well. For this reason, the IP address of the target in the ATS is hard-coded in the System Definitions for both tests.

This code must be built against VS 2020 (not VS 2020R3 which is currently on the build nodes) due to VxWorks support being dropped. The PR build will fail until a build node is reprovisioned.